Fase 2/Evidencias Proyecto/appSmart/requirements/api.txt
# API Requirements - Para escalabilidad multiplataforma
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0  # JWT para móviles
django-cors-headers==4.3.1            # CORS para apps externas
drf-spectacular==0.27.0               # Documentación OpenAPI automática
